Implementation of default software-based key derivation function.
The default software KDF invokes an IKM_HOOK script for initial
keying material and uses the supplied info argument and a
generated salt value to derive a key that can be used to
encrypt other items in EdgeX.

The security of the KDF depends on the confidentiality of
the input key material; thus, IKM_HOOK should use a robust
method to protect the key material at rest. Other mechanisms
should be used to protect the key material at runtime.

kdf-salt.dat is included so that "go test" generates a valid result

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>