Implement encryption of Vault master key by user-supplied key.
User supplied key is obtained by calling a key derivation hook
($KDF_HOOK -persistdir path) and expecting a 32-byte hex-encoded key.
If $KDF_HOOK is undefined, will attempt to locate an execute
'security-default-kdf' on the PATH.  The resulting key is used
for AES-256-GCM encryption of the key shares with a random
nonce for each key share.

The init-resp.json file is modified by removing Keys and
KeysBase64 fields and replacing them with EncryptedKeys and Nonces
fields, and the reverse is done upon decryption.  No decryption
is done if the init-resp.json already contains a KeysBase64 field.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>